+++ src/main/java/frc/robot/commands/archive/ShooterClose.java
@@ -1,4 +1,4 @@
-package frc.robot.commands;
+package frc.robot.commands.archive;
 
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.subsystems.ShooterJaws;
