+++ src/main/java/frc/robot/commands/groups/ScoreTrap.java
@@ -5,8 +5,8 @@ import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.commands.ElevatorLower;
 import frc.robot.commands.Shooter1Out;
 import frc.robot.commands.Shooter2Out;
-import frc.robot.commands.ShooterClose;
 import frc.robot.commands.TriggerOut;
+import frc.robot.commands.archive.ShooterClose;
 import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.ShooterJaws;
 import frc.robot.subsystems.ShooterOne;
