+++ src/main/java/frc/robot/commands/groups/ScoreTrapPrep.java
@@ -1,9 +1,9 @@
 package frc.robot.commands.groups;
 
 import edu.wpi.first.wpilibj2.command.SequentialCommandGroup;
-import frc.robot.commands.AimAmp;
-import frc.robot.commands.ArmExtend;
 import frc.robot.commands.ElevatorRaise;
+import frc.robot.commands.archive.AimAmp;
+import frc.robot.commands.archive.ArmExtend;
 import frc.robot.subsystems.Arm;
 import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.ShooterJaws;
