+++ src/main/java/frc/robot/commands/groups/ShootSpeaker.java
@@ -2,12 +2,12 @@ package frc.robot.commands.groups;
 
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
-import frc.robot.commands.AimSpeaker;
-import frc.robot.commands.ArmRetract;
 import frc.robot.commands.Shooter1Out;
 import frc.robot.commands.Shooter2Out;
-import frc.robot.commands.ShooterClose;
 import frc.robot.commands.TriggerOut;
+import frc.robot.commands.archive.AimSpeaker;
+import frc.robot.commands.archive.ArmRetract;
+import frc.robot.commands.archive.ShooterClose;
 import frc.robot.subsystems.Arm;
 import frc.robot.subsystems.ShooterJaws;
 import frc.robot.subsystems.ShooterOne;
