+++ src/main/java/frc/robot/commands/groups/ShooterIn.java
@@ -3,8 +3,8 @@ package frc.robot.commands.groups;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import frc.robot.commands.Shooter1In;
 import frc.robot.commands.Shooter2In;
-import frc.robot.commands.ShooterOpen;
 import frc.robot.commands.TriggerIn;
+import frc.robot.commands.archive.ShooterOpen;
 import frc.robot.subsystems.ShooterJaws;
 import frc.robot.subsystems.ShooterOne;
 import frc.robot.subsystems.ShooterTrigger;
