+++ src/main/java/frc/robot/commands/groups/ShooterOut.java
@@ -3,7 +3,7 @@ package frc.robot.commands.groups;
 import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import frc.robot.commands.Shooter1Out;
 import frc.robot.commands.Shooter2Out;
-import frc.robot.commands.ShooterClose;
+import frc.robot.commands.archive.ShooterClose;
 import frc.robot.subsystems.ShooterJaws;
 import frc.robot.subsystems.ShooterOne;
 import frc.robot.subsystems.ShooterTwo;
