+++ src/main/java/frc/robot/commands/groups/ShooterOutTrap.java
@@ -4,8 +4,8 @@ import edu.wpi.first.wpilibj2.command.ParallelCommandGroup;
 import edu.wpi.first.wpilibj2.command.WaitCommand;
 import frc.robot.commands.Shooter1Out;
 import frc.robot.commands.Shooter2Out;
-import frc.robot.commands.ShooterClose;
 import frc.robot.commands.TriggerOut;
+import frc.robot.commands.archive.ShooterClose;
 import frc.robot.subsystems.ShooterJaws;
 import frc.robot.subsystems.ShooterOne;
 import frc.robot.subsystems.ShooterTrigger;
